rename LearningPath
